Move internals to separate module, rm unreleased deprecations
