+++ src/deprecated.jl
@@ -341,89 +341,6 @@ function check_complexity(
     )
 end
 
-# --- Cognitive deprecations -----------------------------------------------
-
-function cognitive_complexity(expr_or_code)
-    Base.depwarn(
-        "`cognitive_complexity(x)` is deprecated; use " *
-        "`measure(CognitiveComplexity(), x)`.",
-        :cognitive_complexity,
-    )
-    return measure(CognitiveComplexity(), expr_or_code)
-end
-
-function cognitive_complexity_report(
-    code::AbstractString;
-    max_complexity::Union{Int, Nothing} = nothing,
-)
-    Base.depwarn(
-        "`cognitive_complexity_report` is deprecated; use " *
-        "`measure_report(CognitiveComplexity(), code; max_value=...)`.",
-        :cognitive_complexity_report,
-    )
-    return measure_report(CognitiveComplexity(), code; max_value = max_complexity)
-end
-
-function file_cognitive_complexity(
-    filepath::AbstractString;
-    max_complexity::Union{Int, Nothing} = nothing,
-)
-    Base.depwarn(
-        "`file_cognitive_complexity` is deprecated; use " *
-        "`file_measure(CognitiveComplexity(), path; max_value=...)`.",
-        :file_cognitive_complexity,
-    )
-    return file_measure(CognitiveComplexity(), filepath; max_value = max_complexity)
-end
-
-function directory_cognitive_complexity(
-    dirpath::AbstractString;
-    recursive::Bool = true,
-    max_complexity::Union{Int, Nothing} = nothing,
-)
-    Base.depwarn(
-        "`directory_cognitive_complexity` is deprecated; use " *
-        "`directory_measure(CognitiveComplexity(), dir; ...)`.",
-        :directory_cognitive_complexity,
-    )
-    return directory_measure(
-        CognitiveComplexity(),
-        dirpath;
-        recursive = recursive,
-        max_value = max_complexity,
-    )
-end
-
-function package_cognitive_complexity(
-    pkg;
-    max_complexity::Union{Int, Nothing} = nothing,
-)
-    Base.depwarn(
-        "`package_cognitive_complexity` is deprecated; use " *
-        "`package_measure(CognitiveComplexity(), pkg; max_value=...)`.",
-        :package_cognitive_complexity,
-    )
-    return package_measure(CognitiveComplexity(), pkg; max_value = max_complexity)
-end
-
-function check_cognitive_complexity(
-    path_or_pkg;
-    max_complexity::Int = default_max_value(CognitiveComplexity()),
-    throw_on_violation::Bool = true,
-)
-    Base.depwarn(
-        "`check_cognitive_complexity` is deprecated; use " *
-        "`check_measure(CognitiveComplexity(), path_or_pkg; max_value=...)`.",
-        :check_cognitive_complexity,
-    )
-    return check_measure(
-        CognitiveComplexity(),
-        path_or_pkg;
-        max_value = max_complexity,
-        throw_on_violation = throw_on_violation,
-    )
-end
-
 # --- Argument-count deprecations ------------------------------------------
 
 function argument_count_report(
